Move BigDecimal conversion to conversions crate
When I initially refactored that crate out I didn't want to move these
but now I believe they make sense.
