+++ crates/number_conversions/Cargo.toml
@@ -5,5 +5,6 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
+bigdecimal = "0.3"
 num = "0.4"
 primitive-types = { version = "0.10" }
