+++ crates/orderbook/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod account_balances;
 pub mod api;
 pub mod arguments;
-pub mod conversions;
 pub mod database;
 pub mod event_updater;
 pub mod fee_subsidy;
